refactor(open-with): split discovery modules
